backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "mustafairshaidd",email = "mustafa-irshaid_2001@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "httpx (>=0.27.0,<1.0.0)",
    "strawberry-graphql (>=0.217.0,<1.0.0)",
    "sqlmodel (>=0.0.14,<1.0.0)",
    "psycopg2-binary (>=2.9.9,<3.0.0)",
    "alembic (>=1.13.1,<2.0.0)",
]

[tool.poetry]
packages = [{include = "src"}]


[tool.poetry.group.dev.dependencies]
uvicorn = "^0.35.0"
pytest = "^8.0.0"
pytest-cov = "^5.0.0"
flake8 = "^7.0.0"
bandit = "^1.7.0"
safety = "^2.3.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
